Dark Mode
#comment: Dark mode added. Can search logs. Other improvements.

#duration: 3.5

+++ TourPlanner.Model/Log/LogData.cs
@@ -12,7 +12,7 @@ namespace TourPlanner.Model.Log
 
         public string LogDate { get; set; }
 
-        public double LogDistance { get; set; }
+        public int LogDistance { get; set; }
 
         public string LogTotalTime { get; set; }
